#248 harness-test refactoring, tweaks
